第三次精简代码，并加入停车场数据显示 	modified:   app/src/main/java/com/run/sg/amap3d/basic/UiSettingsActivity.java 	new file:   app/src/main/java/com/run/sg/db/ParkDBHelper.java 	new file:   app/src/main/java/com/run/sg/db/ParkInfoTable.java 	new file:   app/src/main/java/com/run/sg/db/RecentlyGoTable.java 	modified:   app/src/main/java/com/run/sg/model/SGParkingLotItemData.java 	modified:   app/src/main/java/com/run/sg/neighbor/SGNeighborActivity.java 	modified:   app/src/main/java/com/run/sg/neighbor/SGSortDistanceFragment.java 	modified:   app/src/main/java/com/run/sg/neighbor/SGSortListAdapter.java 	modified:   app/src/main/java/com/run/sg/neighbor/SGSortPriceFragment.java 	modified:   app/src/main/java/com/run/sg/search/SGMainSearchActivity.java

+++ app/src/main/java/com/run/sg/model/SGParkingLotItemData.java
@@ -7,11 +7,11 @@ public class SGParkingLotItemData {
 
     private String mName;
     private String mAddress;
-    private int mPrice;
+    private double mPrice;
     private double mDistance;
     private int mVacancy;
 
-    public SGParkingLotItemData(String name,String address,int price,double distance,int vacancy){
+    public SGParkingLotItemData(String name, String address, double price, double distance, int vacancy){
         this.mName = name;
         this.mAddress = address;
         this.mPrice = price;
@@ -27,7 +27,7 @@ public class SGParkingLotItemData {
         return mAddress;
     }
 
-    public int getPrice(){
+    public double getPrice(){
         return mPrice;
     }
